+++ workitems/src/robocorp/workitems/__init__.py
@@ -1,4 +1,5 @@
 import logging
+from pathlib import Path
 from typing import Callable, Optional, Union, cast
 
 from robocorp.tasks import get_current_task, task_cache
@@ -11,7 +12,7 @@ from ._exceptions import (
     EmptyQueue,
     to_exception_type,
 )
-from ._types import ExceptionType, JSONType, State
+from ._types import ExceptionType, JSONType, PathType, State
 from ._workitem import Input, Output
 
 __version__ = "1.4.3"
@@ -139,7 +140,7 @@ class Outputs:
     def create(
         self,
         payload: Optional[JSONType] = None,
-        files: Optional[Union[str, list[str]]] = None,
+        files: Optional[Union[PathType, list[PathType]]] = None,
         save: bool = True,
     ) -> Output:
         """Create a new output work item, which can have both a JSON
@@ -161,7 +162,7 @@ class Outputs:
             item.payload = payload
 
         if files is not None:
-            if isinstance(files, str):
+            if isinstance(files, (Path, str)):
                 item.add_files(pattern=files)
             else:
                 for path in files: